Stop using implicit casts.
C++ implicit cast semantics don't match precisely what we need.

+++ glretrace_glx.cpp
@@ -73,7 +73,7 @@ static void retrace_glXMakeCurrent(Trace::Call &call) {
         }
     }
 
-    glws::Drawable *new_drawable = getDrawable(static_cast<unsigned long>(call.arg(1)));
+    glws::Drawable *new_drawable = getDrawable(call.arg(1).toUInt());
     glws::Context *new_context = context_map[call.arg(2).toPointer()];
 
     bool result = ws->makeCurrent(new_drawable, new_context);
@@ -180,7 +180,7 @@ static void retrace_glXMakeContextCurrent(Trace::Call &call) {
         }
     }
 
-    glws::Drawable *new_drawable = getDrawable(static_cast<unsigned long>(call.arg(1)));
+    glws::Drawable *new_drawable = getDrawable(call.arg(1).toUInt());
     glws::Context *new_context = context_map[call.arg(3).toPointer()];
 
     bool result = ws->makeCurrent(new_drawable, new_context);
